+++ components/ChatLine.tsx
@@ -1,11 +1,5 @@
-// import clsx from 'clsx'
-import Balancer from 'react-wrap-balancer';
-
-// wrap Balancer to remove type errors :( - @TODO - fix this ugly hack
-const BalancerWrapper = (props: any) => <Balancer {...props} />;
-
 export type Message = {
-  who: 'bot' | 'user' | undefined;
+  who: "bot" | "user" | undefined;
   message?: string;
 };
 
@@ -33,35 +27,37 @@ export const LoadingChatLine = () => (
 
 // util helper to convert new lines to <br /> tags
 const convertNewLines = (text: string) =>
-  text.split('\n').map((line, i) => (
+  text.split("\n").map((line, i) => (
     <span key={i}>
       {line}
       <br />
     </span>
   ));
 
-export function ChatLine({ who = 'bot', message }: Message) {
+export function ChatLine({ who = "bot", message }: Message) {
   if (!message) {
     return null;
   }
   const formatteMessage = convertNewLines(message);
 
   return (
-    <div className={who != 'bot' ? 'float-right clear-both' : 'float-left clear-both'}>
-      <BalancerWrapper>
-        <div className="float-right mb-5 rounded-lg bg-white px-4 py-5 shadow-lg ring-1 ring-zinc-100 sm:px-6">
-          <div className="flex space-x-3">
-            <div className="flex-1 gap-4">
-              <p className="font-large text-xxl text-gray-900">
-                <a href="#" className="hover:underline">
-                  {who == 'bot' ? 'AI' : 'You'}
-                </a>
-              </p>
-              <p className={'text-gray-400'}>{formatteMessage}</p>
-            </div>
+    <div
+      className={
+        who != "bot" ? "float-right clear-both" : "float-left clear-both"
+      }
+    >
+      <div className="float-right mb-5 rounded-lg bg-white px-4 py-5 shadow-lg ring-1 ring-zinc-100 sm:px-6">
+        <div className="flex space-x-3">
+          <div className="flex-1 gap-4">
+            <p className="font-large text-xxl text-gray-900">
+              <a href="#" className="hover:underline">
+                {who == "bot" ? "AI" : "You"}
+              </a>
+            </p>
+            <p className={"text-gray-400"}>{formatteMessage}</p>
           </div>
         </div>
-      </BalancerWrapper>
+      </div>
     </div>
   );
 }
